+++ CHANGELOG.md
@@ -34,6 +34,7 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
 - Save light client updates and bootstraps in DB.
 - Added more comprehensive tests for `BlockToLightClientHeader`. [PR](https://github.com/prysmaticlabs/prysm/pull/14699)
 - Added an error field to log `Finished building block`.
+- Implemented a new `EmptyExecutionPayloadHeader` function.
 
 ### Changed
 
@@ -78,6 +79,7 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
 - Check kzg commitments align with blobs and proofs for beacon api end point.
 - Revert "Proposer checks gas limit before accepting builder's bid".
 - Updated quic-go to v0.48.2 .
+- Process light client finality updates only for new finalized epochs instead of doing it for every block.
 
 ### Deprecated